+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Allow sending a list of applications to delete when sending a DELETE to /applications
 * driver: Add timestamps to tasks
 * Return applications using a template when querying the /templatesets endpoint
+* Add support for query IPAM (IP Address Management) providers from templates
 
 ## Fixed
 * GUI: Fix task order in the Deploy Log when using AS3 versions 3.26.0 and 3.27.0